Hice responsive la pagina, ademas de separar por componentes itemCarrito
